% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cbind_embedding_similarity}
\alias{cbind_embedding_similarity}
\title{Add similarities of your text to the labels of a Starspace model}
\usage{
cbind_embedding_similarity(object, newdata, type = c("cosine", "dot"))
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}

\item{newdata}{a data frame with columns \code{doc_id} and \code{text} indicating the text for which you want to get the similarity to the labels of the Starspace model}

\item{type}{either 'cosine' or 'dot'. If 'dot', uses inner-product based similarity, if 'cosine', returns cosine similarity.
Passed on to \code{\link{embedding_similarity}}}
}
\value{
The data frame \code{newdata} where columns are added - one for each label indicating the similarity between the text and each of the labels.
}
\description{
Add similarities of your text to the labels of a Starspace model. Similarities are computed with \code{\link{embedding_similarity}}
}
\examples{
data(dekamer, package = "ruimtehol")
dekamer$text <- gsub("\\\\.([[:digit:]]+)\\\\.", ". \\\\1.", x = dekamer$question)
dekamer$text <- strsplit(dekamer$text, "\\\\W")
dekamer$text <- lapply(dekamer$text, FUN = function(x) setdiff(x, ""))
dekamer$text <- sapply(dekamer$text, 
                       FUN = function(x) paste(x, collapse = " "))

idx <- sample(nrow(dekamer), size = round(nrow(dekamer) * 0.9))
traindata <- dekamer[idx, ]
testdata <- dekamer[-idx, ]
model <- embed_tagspace(x = traindata$text, 
                        y = traindata$question_theme_main, 
                        early_stopping = 0.8,
                        dim = 10, minCount = 5)
scores <- cbind_embedding_similarity(model, testdata)
}
