% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dekamer_theme_terminology}
\alias{dekamer_theme_terminology}
\title{Dataset containing relevant terminology for each theme of the \code{dekamer} dataset}
\description{
Dataset containing relevant terminology for each theme of the \code{\link{dekamer}} dataset

The dataset contains the following information: 
\itemize{
\item theme: a theme, corresponding to the \code{question_theme_main} field in the \code{\link{dekamer}} dataset
\item term: a word which describes the \code{theme}
\item n: a measure of information indicating how relevant the term is (frequency of occurrence)
}
}
\examples{
data(dekamer_theme_terminology)
str(dekamer_theme_terminology)
}
