% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-all-the-things.R
\name{embed_wordspace}
\alias{embed_wordspace}
\title{Build a Starspace model which calculates word embeddings}
\usage{
embed_wordspace(x, model = "wordspace.bin", ...)
}
\arguments{
\item{x}{a character vector of text where tokens are separated by spaces}

\item{model}{name of the model which will be saved, passed on to \code{\link{starspace}}}

\item{...}{further arguments passed on to \code{\link{starspace}} except file, trainMode and fileFormat}
}
\value{
an object of class \code{textspace} as returned by \code{\link{starspace}}.
}
\description{
Build a Starspace model which calculates word embeddings
}
\examples{
library(udpipe)
library(tokenizers)
data(brussels_reviews, package = "udpipe")
x <- subset(brussels_reviews, language == "nl")
x <- tokenize_words(x$feedback)
x <- sapply(x, FUN = function(x) paste(x, collapse = " "))

model <- embed_wordspace(x, dim = 15, ws = 7, epoch = 5, minCount = 5, ngrams = 1)
wordvectors <- as.matrix(model)

mostsimilar <- embedding_similarity(wordvectors, wordvectors["weekend", ])
head(sort(mostsimilar[, 1], decreasing = TRUE), 10)
mostsimilar <- embedding_similarity(wordvectors, wordvectors["vriendelijk", ])
head(sort(mostsimilar[, 1], decreasing = TRUE), 10)
mostsimilar <- embedding_similarity(wordvectors, wordvectors["grote", ])
head(sort(mostsimilar[, 1], decreasing = TRUE), 10)
}
