% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{embedding_similarity}
\alias{embedding_similarity}
\title{Cosine and Inner product based similarity}
\usage{
embedding_similarity(x, y, type = c("cosine", "dot"))
}
\arguments{
\item{x}{a matrix with embeddings}

\item{y}{a matrix with embeddings}

\item{type}{either 'cosine' or 'dot'. If 'dot', returns inner-product based similarity, if 'cosine', returns cosine similarity}
}
\value{
similarity matrix between the rows of \code{x} and the rows of \code{y}. 
The similarity between row i of \code{x} and row j of \code{y} is found in cell \code{[i, j]} of the returned similarity matrix
}
\description{
Cosine and Inner product based similarity
}
\examples{
x <- matrix(rnorm(6), nrow = 2, ncol = 3)
rownames(x) <- c("word1", "word2")
y <- matrix(rnorm(15), nrow = 5, ncol = 3)
rownames(y) <- c("term1", "term2", "term3", "term4", "term5")

embedding_similarity(x, y, type = "cosine")
embedding_similarity(x, y, type = "dot")
}
