% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{predict.textspace}
\alias{predict.textspace}
\title{Predict using a Starspace model}
\usage{
\method{predict}{textspace}(object, newdata, sep = " ", basedoc, ...)
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}

\item{newdata}{a character string of length 1}

\item{sep}{character string used to split \code{newdata} using boost::split}

\item{basedoc}{optional, the path to a file in labelDoc format, containing basedocs which are set of possible things to predict, if different than 
the ones from the training data}

\item{...}{not used}
}
\value{
a list with elements input and a data.frame called prediction which has columns called label and similarity
}
\description{
Predict using a Starspace model
}
