% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{starspace_dictionary}
\alias{starspace_dictionary}
\title{Get the dictionary of a Starspace model}
\usage{
starspace_dictionary(object)
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}
}
\value{
a list with elements 
\enumerate{
\item{ntokens: }{The number of tokens in the data}
\item{nwords: }{The number of words which are part of the dictionary}
\item{nlabels: }{The number of labels which are part of the dictionary}
\item{labels: }{A character vector with the labels}
\item{dictionary_size: }{The size of the dictionary (nwords + nlabels)}
\item{dictionary: }{A data.frame with all the words and labels from the dictionary. This data.frame has columns term, is_word and is_label indicating
for each term if it is a word or a label}
}
}
\description{
Get the dictionary of a Starspace model
}
