% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{starspace_embedding}
\alias{starspace_embedding}
\title{Get the document embeddings}
\usage{
starspace_embedding(object, x, type = "document")
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}

\item{x}{character string with text}

\item{type}{the type of embedding requested. Currently only 'document' is possible}
}
\value{
a matrix
}
\description{
Get the document embeddings
}
