% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{starspace_save_model}
\alias{starspace_save_model}
\title{Save a starspace model as a binary or tab-delimited TSV file}
\usage{
starspace_save_model(object, file = "textspace.tsv", method = c("binary",
  "tsv-starspace", "tsv-data.table"))
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}

\item{file}{character string with the path to the file where to save the model, in case as_tsv is set to \code{TRUE}}

\item{method}{character indicating the method of saving. Possible values are 'binary', 'tsv-starspace', 'tsv-data.table'. Defaults to 'binary'.
\enumerate{
\item{The first method: \code{'binary'} saves the model as a binary file using the Starspace methods}
\item{The second method: \code{'tsv-starspace'} saves the model as a tab-delimited flat file using the Starspace methods}
\item{The third method: \code{'tsv-data.table'} saves the model as a tab-delimited flat file using the fast data.table fread function}
}}
}
\value{
invisibly, the character string with the file of the saved object
}
\description{
Save a starspace model as a binary or a tab-delimited TSV file
}
