% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-all-the-things.R
\name{textspace_classify}
\alias{textspace_classify}
\title{Build a Starspace model to be used for classification purposes}
\usage{
textspace_classify(x, y, model = "textspace_classifier.bin", ...)
}
\arguments{
\item{x}{a character vector of text where tokens are separated by spaces}

\item{y}{a character vector of classes to predict or a list with the same length of \code{x} with several classes for each respective element of \code{x}}

\item{model}{name of the model which will be saved, passed on to \code{\link{starspace}}}

\item{...}{further arguments passed on to \code{\link{starspace}}}
}
\value{
an object of class \code{textspace} as returned by \code{\link{starspace}}.
}
\description{
Build a Starspace model to be used for classification purposes
}
\examples{
library(udpipe)
library(tokenizers)
data(brussels_listings, package = "udpipe")
x <- tokenize_words(brussels_listings$name)
x <- sapply(x, FUN=txt_collapse)
model <- textspace_classify(x = x, y = brussels_listings$room_type, 
                            dim = 10, minCount = 5)
predict(model, "room close to centre gare du midi")
starspace_embedding(model, "room close to centre gare du midi")
}
