% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{predict.textspace}
\alias{predict.textspace}
\title{Predict using a Starspace model}
\usage{
\method{predict}{textspace}(object, newdata, k = 5L, sep = " ", basedoc,
  ...)
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}

\item{newdata}{a character string of length 1}

\item{k}{integer with the number of predictions to make. Defaults to 5.}

\item{sep}{character string used to split \code{newdata} using boost::split}

\item{basedoc}{optional, either a character vector of possible elements to predict or 
the path to a file in labelDoc format, containing basedocs which are set of possible things to predict, if different than 
the ones from the training data}

\item{...}{not used}
}
\value{
a list with elements 
\enumerate{
\item input: the character string passed on to newdata
\item prediction: data.frame called prediction which has columns called label and similarity indicating the predicted label and the similarity of the input ot the label
\item terms: a list with elements basedoc_index and basedoc_terms indicating the position in basedoc and the terms which are part of the dictionary which are used to find the similarity
}
}
\description{
Predict using a Starspace model
}
