% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{starspace}
\alias{starspace}
\title{Interface to Starspace}
\usage{
starspace(file = "textspace.bin", ...)
}
\arguments{
\item{file}{the path to where the model file will be save. Defaults to 'textspace.bin'}

\item{...}{UNDER CONSTRUCTION. Currently just look to the C++ code of the textspace function in the src/rcpp_textspace.cpp 
or type ruimtehol:::textspace_help() for seeing the Starspace options}
}
\value{
an object of class textspace
}
\description{
Interface to Starspace, providing raw access to the textspace functionality. For expert use only.
}
