% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{starspace_load_model}
\alias{starspace_load_model}
\title{Load a Starspace model}
\usage{
starspace_load_model(object)
}
\arguments{
\item{object}{either the path to a Starspace model on disk or an object of class \code{textspace} which you want to reload.}
}
\value{
an object of class textspace
}
\description{
Load a Starspace model
}
