% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed-all-the-things.R
\name{starspace_save_model}
\alias{starspace_save_model}
\title{Save a starspace model as a tab-delimited TSV file}
\usage{
starspace_save_model(object, file = "textspace.tsv")
}
\arguments{
\item{object}{an object of class \code{textspace} as returned by \code{\link{starspace}} or \code{\link{starspace_load_model}}}

\item{file}{character string with the path to the file where to save the model}
}
\value{
invisibly, the \code{file} with the location of the TSV file
}
\description{
Save a starspace model as a tab-delimited TSV file
}
